Begin Material Design Template
Began work on creating a meterial design template.  The top app bar was mostly done, the bottom app bar is partially done.  Several icons have been imported for use.  Began working on scripts that would make the title change and fade in and out when states change.  This would be a different way that the scene change method.